requirements.txt
streamlit==1.30.0
tensorflow==2.20.0
numpy==1.25.2
Pillow==10.0.0